Datepicker removed. frame-budget and frame-new-note created. The value of the budget was changed from int to double.

+++ src/main/java/com/yorix/autometer/controller/ParamController.java
@@ -21,7 +21,7 @@ public class ParamController {
     }
 
     @GetMapping("{name}/")
-    public int get(@PathVariable("name") String name) {
+    public double get(@PathVariable("name") String name) {
         return paramService.read(name);
     }
